Functionality to edit your profile information
A new modal component was created that will allow you to see your profile information and edit it.
The updated information is correctly uploaded to the database and is displayed on your profile screen.
New fields to edit your information and a new button is shown once your account is in Host mode.