Optimization and bug fixes
Small changes were made to the calendar code to make it more stable and faster, as well as fixing some bugs in the calendar when selecting dates.
Added a feature to the calendar that disables dates that have already registered a reservation, this to avoid multiple reservations on the same dates.