Delete all property data
A new call to the database was implemented to not only delete the property from the 'properties' collection but also delete the references to that property within the user's collection.
Also now deleting the property also deletes the images of this property that were hosted in FireStorage.
A new reference field was added in the user's reservations to facilitate the deletion of these.